Split Grammar into 2 to separate interface
