+++ index.ts
@@ -32,6 +32,8 @@ export * from "./lib/Microgrammar";
 
 export * from "./lib/Grammar";
 
+export * from "./lib/microgrammarConstruction";
+
 export * from "./lib/Ops";
 
 export { isPatternMatch } from "./lib/PatternMatch";
